requirements.txt
pandas==2.3.2
praw==7.8.1
pyarrow==21.0.0
python-dotenv==1.1.1
requests==2.32.5
tqdm==4.67.1
